+++ src/main/java/dev/muon/otherworldapoth/mixin/apoth/loot/AffixLootPoolEntryMixin.java
@@ -1,6 +1,6 @@
-package dev.muon.otherworldapoth.mixin.apoth;
+package dev.muon.otherworldapoth.mixin.apoth.loot;
 
-import dev.shadowsoffire.apotheosis.adventure.loot.GemLootPoolEntry;
+import dev.shadowsoffire.apotheosis.adventure.loot.AffixLootPoolEntry;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootContext;
 import org.spongepowered.asm.mixin.Mixin;
@@ -10,10 +10,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.function.Consumer;
 
-@Mixin(value = GemLootPoolEntry.class, remap = false)
-public class GemLootPoolEntryMixin {
+@Mixin(value = AffixLootPoolEntry.class, remap = false)
+public class AffixLootPoolEntryMixin {
     @Inject(method = "createItemStack", at = @At("HEAD"), cancellable = true, remap = true)
-    private void cancelGemInjection(Consumer<ItemStack> list, LootContext ctx, CallbackInfo ci) {
+    private void cancelAffixInjection(Consumer<ItemStack> list, LootContext ctx, CallbackInfo ci) {
         // TODO: We might want to use this system in some cases, instead of disabling it entirely
         ci.cancel();
     }